push/pyproject.toml
[project]
name = "push"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "pillow>=11.3.0",
]
